+++ src/resource_clients/run.ts
@@ -142,6 +142,7 @@ export class RunClient extends ResourceClient {
                 timeout: ow.optional.number,
                 maxItems: ow.optional.number,
                 maxTotalChargeUsd: ow.optional.number,
+                restartOnError: ow.optional.boolean,
             }),
         );
 
@@ -320,6 +321,7 @@ export interface RunResurrectOptions {
     timeout?: number;
     maxItems?: number;
     maxTotalChargeUsd?: number;
+    restartOnError?: boolean;
 }
 
 export interface RunChargeOptions {